Fix pointer math casting
* Apparently pointer subtraction will return the result as ptrdiff_t, and already divided by the type size. So fix all cases where this is done and switch to a more C-like approach. The C++ approach is kind of unexpected, and has more cases of undefined behaviour.

+++ src/openlrr/engine/core/ListSet.hpp
@@ -88,7 +88,8 @@ constexpr const size_t CountOfList(size_t listIndex)
 template <typename TItem>
 constexpr const size_t IndexOfInList(const TItem* list, const TItem* item)
 {
-	return (static_cast<size_t>(item - list) / sizeof(TItem));
+	/// POINTER MATH: Cast to ptrdiff_t before subtraction to avoid C++ pointer math conversion.
+	return (static_cast<size_t>((ptrdiff_t)item - (ptrdiff_t)list) / sizeof(TItem));
 }
 
 /**
